app_numeros/requirements.txt

flask==2.0.1
tensorflow==2.16.1  
pillow==10.3.0
numpy==1.26.4
gunicorn==21.2.0